Added changes code refinement and invitation code

+++ src/main/java/com/board/notification/dao/GroupRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.board.notification.model.Groups;
 
-public interface GroupRepo extends CrudRepository<Groups, Long> {
+public interface GroupRepo extends CrudRepository<Groups, Integer> {
 	public Groups findByGroupName(String groupName);
 }
